fix some more 32/64bit incompatibilities

+++ libpcsxcore/coff.h
@@ -25,9 +25,9 @@
 struct external_filehdr {
 	unsigned short f_magic;		/* magic number			*/
 	unsigned short f_nscns;		/* number of sections		*/
-	unsigned long f_timdat;	/* time & date stamp		*/
-	unsigned long f_symptr;	/* file pointer to symtab	*/
-	unsigned long f_nsyms;		/* number of symtab entries	*/
+	unsigned int f_timdat;	/* time & date stamp		*/
+	unsigned int f_symptr;	/* file pointer to symtab	*/
+	unsigned int f_nsyms;		/* number of symtab entries	*/
 	unsigned short f_opthdr;	/* sizeof(optional hdr)		*/
 	unsigned short f_flags;		/* flags			*/
 };
